requirements.txt
databases==0.7.0
fastapi==0.94.0
#jose==1.0.0
PyJwt
passlib==1.7.4
pydantic==1.10.5
python_jose==3.3.0
SQLAlchemy==1.4.46
uvicorn==0.15.0
asyncpg